requirements.txt
# requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
selenium>=4.8.0          # For JavaScript-heavy sites
schedule>=1.2.0          # For scheduling
python-dotenv>=1.0.0
sqlalchemy>=2.0.0        # Database ORM
flask>=2.3.0             # Optional web interface
pandas>=2.0.0            # Data manipulation
tabulate>=0.9.0          # For CLI table formatting